Day Generator
I took out the build frequency function because it wasn't working and didn't do anything and then I added the day generator I was thinking about before and it worked great. It now generates a message object every minute that the call frequency is set higher than the frequency property of the client object and I also started working on the Answer Service object which is ultimately going to store all the saved variables pertaining to the game state.